+++ tests/index.test.ts
@@ -23,7 +23,7 @@ describe('instantiate client', () => {
     const client = new Gentrace({
       baseURL: 'http://localhost:5000/',
       defaultHeaders: { 'X-My-Default-Header': '2' },
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
     });
 
     test('they are used in the request', () => {
@@ -87,14 +87,14 @@ describe('instantiate client', () => {
         error: jest.fn(),
       };
 
-      const client = new Gentrace({ logger: logger, logLevel: 'debug', bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, logLevel: 'debug', apiKey: 'My API Key' });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).toHaveBeenCalled();
     });
 
     test('default logLevel is warn', async () => {
-      const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ apiKey: 'My API Key' });
       expect(client.logLevel).toBe('warn');
     });
 
@@ -107,7 +107,7 @@ describe('instantiate client', () => {
         error: jest.fn(),
       };
 
-      const client = new Gentrace({ logger: logger, logLevel: 'info', bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, logLevel: 'info', apiKey: 'My API Key' });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).not.toHaveBeenCalled();
@@ -123,7 +123,7 @@ describe('instantiate client', () => {
       };
 
       process.env['GENTRACE_LOG'] = 'debug';
-      const client = new Gentrace({ logger: logger, bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, apiKey: 'My API Key' });
       expect(client.logLevel).toBe('debug');
 
       await forceAPIResponseForClient(client);
@@ -140,7 +140,7 @@ describe('instantiate client', () => {
       };
 
       process.env['GENTRACE_LOG'] = 'not a log level';
-      const client = new Gentrace({ logger: logger, bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, apiKey: 'My API Key' });
       expect(client.logLevel).toBe('warn');
       expect(warnMock).toHaveBeenCalledWith(
         'process.env[\'GENTRACE_LOG\'] was set to "not a log level", expected one of ["off","error","warn","info","debug"]',
@@ -157,7 +157,7 @@ describe('instantiate client', () => {
       };
 
       process.env['GENTRACE_LOG'] = 'debug';
-      const client = new Gentrace({ logger: logger, logLevel: 'off', bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, logLevel: 'off', apiKey: 'My API Key' });
 
       await forceAPIResponseForClient(client);
       expect(debugMock).not.toHaveBeenCalled();
@@ -173,7 +173,7 @@ describe('instantiate client', () => {
       };
 
       process.env['GENTRACE_LOG'] = 'not a log level';
-      const client = new Gentrace({ logger: logger, logLevel: 'debug', bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ logger: logger, logLevel: 'debug', apiKey: 'My API Key' });
       expect(client.logLevel).toBe('debug');
       expect(warnMock).not.toHaveBeenCalled();
     });
@@ -184,7 +184,7 @@ describe('instantiate client', () => {
       const client = new Gentrace({
         baseURL: 'http://localhost:5000/',
         defaultQuery: { apiVersion: 'foo' },
-        bearerToken: 'My Bearer Token',
+        apiKey: 'My API Key',
       });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/foo?apiVersion=foo');
     });
@@ -193,7 +193,7 @@ describe('instantiate client', () => {
       const client = new Gentrace({
         baseURL: 'http://localhost:5000/',
         defaultQuery: { apiVersion: 'foo', hello: 'world' },
-        bearerToken: 'My Bearer Token',
+        apiKey: 'My API Key',
       });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/foo?apiVersion=foo&hello=world');
     });
@@ -202,7 +202,7 @@ describe('instantiate client', () => {
       const client = new Gentrace({
         baseURL: 'http://localhost:5000/',
         defaultQuery: { hello: 'world' },
-        bearerToken: 'My Bearer Token',
+        apiKey: 'My API Key',
       });
       expect(client.buildURL('/foo', { hello: undefined })).toEqual('http://localhost:5000/foo');
     });
@@ -211,7 +211,7 @@ describe('instantiate client', () => {
   test('custom fetch', async () => {
     const client = new Gentrace({
       baseURL: 'http://localhost:5000/',
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
       fetch: (url) => {
         return Promise.resolve(
           new Response(JSON.stringify({ url, custom: true }), {
@@ -229,7 +229,7 @@ describe('instantiate client', () => {
     // make sure the global fetch type is assignable to our Fetch type
     const client = new Gentrace({
       baseURL: 'http://localhost:5000/',
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
       fetch: defaultFetch,
     });
   });
@@ -237,7 +237,7 @@ describe('instantiate client', () => {
   test('custom signal', async () => {
     const client = new Gentrace({
       baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
       fetch: (...args) => {
         return new Promise((resolve, reject) =>
           setTimeout(
@@ -269,7 +269,7 @@ describe('instantiate client', () => {
 
     const client = new Gentrace({
       baseURL: 'http://localhost:5000/',
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
       fetch: testFetch,
     });
 
@@ -279,18 +279,12 @@ describe('instantiate client', () => {
 
   describe('baseUrl', () => {
     test('trailing slash', () => {
-      const client = new Gentrace({
-        baseURL: 'http://localhost:5000/custom/path/',
-        bearerToken: 'My Bearer Token',
-      });
+      const client = new Gentrace({ baseURL: 'http://localhost:5000/custom/path/', apiKey: 'My API Key' });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/custom/path/foo');
     });
 
     test('no trailing slash', () => {
-      const client = new Gentrace({
-        baseURL: 'http://localhost:5000/custom/path',
-        bearerToken: 'My Bearer Token',
-      });
+      const client = new Gentrace({ baseURL: 'http://localhost:5000/custom/path', apiKey: 'My API Key' });
       expect(client.buildURL('/foo', null)).toEqual('http://localhost:5000/custom/path/foo');
     });
 
@@ -299,45 +293,41 @@ describe('instantiate client', () => {
     });
 
     test('explicit option', () => {
-      const client = new Gentrace({ baseURL: 'https://example.com', bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ baseURL: 'https://example.com', apiKey: 'My API Key' });
       expect(client.baseURL).toEqual('https://example.com');
     });
 
     test('env variable', () => {
       process.env['GENTRACE_BASE_URL'] = 'https://example.com/from_env';
-      const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ apiKey: 'My API Key' });
       expect(client.baseURL).toEqual('https://example.com/from_env');
     });
 
     test('empty env variable', () => {
       process.env['GENTRACE_BASE_URL'] = ''; // empty
-      const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ apiKey: 'My API Key' });
       expect(client.baseURL).toEqual('https://gentrace.ai/api');
     });
 
     test('blank env variable', () => {
       process.env['GENTRACE_BASE_URL'] = '  '; // blank
-      const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+      const client = new Gentrace({ apiKey: 'My API Key' });
       expect(client.baseURL).toEqual('https://gentrace.ai/api');
     });
   });
 
   test('maxRetries option is correctly set', () => {
-    const client = new Gentrace({ maxRetries: 4, bearerToken: 'My Bearer Token' });
+    const client = new Gentrace({ maxRetries: 4, apiKey: 'My API Key' });
     expect(client.maxRetries).toEqual(4);
 
     // default
-    const client2 = new Gentrace({ bearerToken: 'My Bearer Token' });
+    const client2 = new Gentrace({ apiKey: 'My API Key' });
     expect(client2.maxRetries).toEqual(2);
   });
 
   describe('withOptions', () => {
     test('creates a new client with overridden options', () => {
-      const client = new Gentrace({
-        baseURL: 'http://localhost:5000/',
-        maxRetries: 3,
-        bearerToken: 'My Bearer Token',
-      });
+      const client = new Gentrace({ baseURL: 'http://localhost:5000/', maxRetries: 3, apiKey: 'My API Key' });
 
       const newClient = client.withOptions({
         maxRetries: 5,
@@ -362,7 +352,7 @@ describe('instantiate client', () => {
         baseURL: 'http://localhost:5000/',
         defaultHeaders: { 'X-Test-Header': 'test-value' },
         defaultQuery: { 'test-param': 'test-value' },
-        bearerToken: 'My Bearer Token',
+        apiKey: 'My API Key',
       });
 
       const newClient = client.withOptions({
@@ -377,11 +367,7 @@ describe('instantiate client', () => {
     });
 
     test('respects runtime property changes when creating new client', () => {
-      const client = new Gentrace({
-        baseURL: 'http://localhost:5000/',
-        timeout: 1000,
-        bearerToken: 'My Bearer Token',
-      });
+      const client = new Gentrace({ baseURL: 'http://localhost:5000/', timeout: 1000, apiKey: 'My API Key' });
 
       // Modify the client properties directly after creation
       client.baseURL = 'http://localhost:6000/';
@@ -409,21 +395,21 @@ describe('instantiate client', () => {
 
   test('with environment variable arguments', () => {
     // set options via env var
-    process.env['GENTRACE_API_KEY'] = 'My Bearer Token';
+    process.env['GENTRACE_API_KEY'] = 'My API Key';
     const client = new Gentrace();
-    expect(client.bearerToken).toBe('My Bearer Token');
+    expect(client.apiKey).toBe('My API Key');
   });
 
   test('with overridden environment variable arguments', () => {
     // set options via env var
-    process.env['GENTRACE_API_KEY'] = 'another My Bearer Token';
-    const client = new Gentrace({ bearerToken: 'My Bearer Token' });
-    expect(client.bearerToken).toBe('My Bearer Token');
+    process.env['GENTRACE_API_KEY'] = 'another My API Key';
+    const client = new Gentrace({ apiKey: 'My API Key' });
+    expect(client.apiKey).toBe('My API Key');
   });
 });
 
 describe('request building', () => {
-  const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+  const client = new Gentrace({ apiKey: 'My API Key' });
 
   describe('custom headers', () => {
     test('handles undefined', () => {
@@ -442,7 +428,7 @@ describe('request building', () => {
 });
 
 describe('default encoder', () => {
-  const client = new Gentrace({ bearerToken: 'My Bearer Token' });
+  const client = new Gentrace({ apiKey: 'My API Key' });
 
   class Serializable {
     toJSON() {
@@ -527,7 +513,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', timeout: 10, fetch: testFetch });
+    const client = new Gentrace({ apiKey: 'My API Key', timeout: 10, fetch: testFetch });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);
@@ -557,7 +543,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', fetch: testFetch, maxRetries: 4 });
+    const client = new Gentrace({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
 
@@ -581,7 +567,7 @@ describe('retries', () => {
       capturedRequest = init;
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', fetch: testFetch, maxRetries: 4 });
+    const client = new Gentrace({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
 
     expect(
       await client.request({
@@ -611,7 +597,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
     const client = new Gentrace({
-      bearerToken: 'My Bearer Token',
+      apiKey: 'My API Key',
       fetch: testFetch,
       maxRetries: 4,
       defaultHeaders: { 'X-Stainless-Retry-Count': null },
@@ -643,7 +629,7 @@ describe('retries', () => {
       capturedRequest = init;
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', fetch: testFetch, maxRetries: 4 });
+    const client = new Gentrace({ apiKey: 'My API Key', fetch: testFetch, maxRetries: 4 });
 
     expect(
       await client.request({
@@ -673,7 +659,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', fetch: testFetch });
+    const client = new Gentrace({ apiKey: 'My API Key', fetch: testFetch });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);
@@ -703,7 +689,7 @@ describe('retries', () => {
       return new Response(JSON.stringify({ a: 1 }), { headers: { 'Content-Type': 'application/json' } });
     };
 
-    const client = new Gentrace({ bearerToken: 'My Bearer Token', fetch: testFetch });
+    const client = new Gentrace({ apiKey: 'My API Key', fetch: testFetch });
 
     expect(await client.request({ path: '/foo', method: 'get' })).toEqual({ a: 1 });
     expect(count).toEqual(2);